requirements.txt
torch
torchvision
opencv-python-headless
craft-text-detector-updated
dateparser
pytesseract
lxml